pyproject.toml
[project]
name = "aether-sre"
version = "0.1.0"
description = "Autonomous AI-SRE Agent for K8s"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langgraph>=0.2.0",
    "pydantic-ai>=0.0.1",
    "kubernetes-asyncio>=30.0.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.26.0",
    "python-dotenv>=1.0.1",
]